Add class for TableOfConents
This will be helpful for my work in another branch too
