+++ website/components/chapter.js
@@ -1,23 +1,9 @@
 import Layout from "./layout";
+import bookContents from "../data/table_of_contents";
 import styles from "./chapter.module.css";
-import tableOfContents from "../table_of_contents";
-
-function findChapterInfo(fileName) {
-  const chapter = tableOfContents.filter(
-    (chapter) => chapter.file === fileName
-  );
-
-  if (chapter.length == 0) {
-    throw new Error("File name not found");
-  } else if (chapter.length > 1) {
-    throw new Error("File name not unique");
-  }
-
-  return chapter[0];
-}
 
 export default function Chapter({ children, fileName }) {
-  const chapter = findChapterInfo(fileName);
+  const chapter = bookContents.getChapterInfo(fileName);
   return (
     <Layout>
       <article>
