--- website/table_of_contents.js
@@ -1,29 +1,0 @@
-const tableOfContents = [
-  {
-    chapterNumber: "0",
-    title: "Introduction",
-    file: "introduction",
-  },
-  {
-    chapterNumber: "1",
-    title: "Linear Regression",
-    file: "linear_regression",
-  },
-  {
-    chapterNumber: "2",
-    title: "Assessing Performance",
-    file: "assessing_performance",
-  },
-  {
-    chapterNumber: "3",
-    title: "Regularization: Ridge",
-    file: "ridge",
-  },
-  {
-    chapterNumber: "99",
-    title: "Tufte (layout example)",
-    file: "tufte",
-  },
-];
-
-export default tableOfContents;
